+++ lambda/api/subscribe_folder/function.py
@@ -1,6 +1,7 @@
 '''Marks the specified folder as subscribed'''
 import json
 from helper import subscribe_folder # pylint: disable=import-error
+from helper import parse_json_body # pylint: disable=import-error
 
 from helper import maintenance_guard # pylint: disable=import-error
 
@@ -8,7 +9,9 @@ from helper import maintenance_guard # pylint: disable=import-error
 @maintenance_guard
 def handler(event, _context):
     '''Marks the specified folder as subscribed'''
-    body = json.loads(event['body'])
+    body, error = parse_json_body(event)
+    if error:
+        return error
     user = event['requestContext']['authorizer']['claims']['cognito:username']
     status = subscribe_folder(body['folder'].replace("/","."), body['host'], user)
     return {